[#46] 화면 선언
